# swagger.yaml
openapi: 3.0.0
info:
  title: E-commerce API
  description: Complete API documentation for E-commerce platform with product categories, brands, models, layout management, role-based access, reCAPTCHA protection, and auto-generated IDs
  version: 4.0.0
  contact:
    name: API Support
    email: support@example.com
servers:
  - url: http://localhost:3001/api
    description: Development server
  - url: https://api.ecommerce.com/v1
    description: Production server
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    CategoryInfo:
      type: object
      properties:
        id:
          type: string
          example: "CAT123456781234"
          description: "Auto-generated category ID (format: CAT + timestamp + random)"
        name:
          type: string
          example: "Processors"
        slug:
          type: string
          example: "processors"
        brands:
          type: array
          items:
            type: string
          example: ["Intel", "AMD"]
          description: "List of brands available in this category"
...
    Product:
      type: object
      properties:
        _id:
          type: string
        productId:
          type: string
          example: "PRD123456781234"
          description: "Auto-generated product ID (format: PRD + timestamp + random)"
        name:
          type: string
          example: "Intel Core i9-13900K"
        heading:
          type: string
          example: "24-Core Desktop Processor"
        description:
          type: string
        category:
          $ref: '#/components/schemas/CategoryInfo'
        brand:
          type: string
          example: "Intel"
          description: "Product brand (required)"
        model:
          type: string
          example: "Core i9-13900K"
          description: "Product model (optional)"
        price:
          type: number
          example: 58999
        cost:
          type: number
          example: 52000
        warranty:
          type: string
          example: "3 Years"
        units:
          type: number
          example: 25
        shippingTime:
          type: string
          example: "2-3 business days"
        shippingCost:
          type: number
          example: 0
        images:
          type: array
          items:
            type: object
            properties:
              url:
                type: string
                example: "https://example.com/image.jpg"
              thumb:
                type: string
                example: "https://example.com/image-thumb.jpg"
              isPrimary:
                type: boolean
        specifications:
          type: object
          additionalProperties:
            type: string
          example: {}
        inStock:
          type: boolean
        soldCount:
          type: number
        ratings:
          type: object
          properties:
            average:
              type: number
              example: 4.5
            count:
              type: number
              example: 128
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    Address:
      type: object
      required:
        - name
        - phone
        - addressLine1
        - city
        - state
        - pincode
      properties:
        _id:
          type: string
        type:
          type: string
          enum: [primary, secondary]
          default: secondary
        name:
          type: string
          example: "John Doe"
        phone:
          type: string
          example: "9876543210"
        addressLine1:
          type: string
          example: "123 Main Street"
        addressLine2:
          type: string
          example: "Apartment 4B"
        city:
          type: string
          example: "Mumbai"
        state:
          type: string
          example: "Maharashtra"
        pincode:
          type: string
          example: "400001"
        country:
          type: string
          default: "India"
        isDefault:
          type: boolean
          default: false

    # ============= Order Schemas (UPDATED with Brand and Model) =============
    Order:
      type: object
      properties:
        _id:
          type: string
        orderId:
          type: string
          example: "ORD123456789"
          description: "Auto-generated order ID"
        user:
          type: object
          properties:
            _id:
              type: string
            name:
              type: string
            email:
              type: string
        userId:
          type: string
        userName:
          type: string
        items:
          type: array
          items:
            type: object
            properties:
              product:
                $ref: '#/components/schemas/Product'
              productId:
                type: string
              name:
                type: string
              brand:
                type: string
              model:
                type: string
              quantity:
                type: number
              price:
                type: number
              total:
                type: number
        shippingAddress:
          $ref: '#/components/schemas/Address'
        billingAddress:
          $ref: '#/components/schemas/Address'
        subtotal:
          type: number
        shippingCost:
          type: number
        tax:
          type: number
        total:
          type: number
        status:
          type: string
          enum: [pending, accepted, ready_to_ship, picked_up, in_transit, delivered, cancelled, refunded]
        payment:
          type: object
          properties:
            razorpayOrderId:
              type: string
            razorpayPaymentId:
              type: string
            razorpaySignature:
              type: string
            method:
              type: string
              enum: [razorpay, cod]
            status:
              type: string
              enum: [pending, completed, failed, refunded]
            amount:
              type: number
            transactionId:
              type: string
            refundId:
              type: string
            refundedAt:
              type: string
              format: date-time
        statusHistory:
          type: array
          items:
            type: object
            properties:
              status:
                type: string
              date:
                type: string
                format: date-time
              note:
                type: string
              trackingNumber:
                type: string
        trackingNumber:
          type: string
        estimatedDelivery:
          type: string
          format: date-time
        deliveredAt:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
...
paths:
...
  /users/orders:
    get:
      tags: [Users]
      summary: Get own orders
      security:
        - bearerAuth: []
      responses:
        200:
          description: List of user orders
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  count:
                    type: number
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Order'
  /users/orders/{orderId}:
    get:
      tags: [Users]
      summary: Get specific own order
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: orderId
          required: true
          schema:
            type: string
      responses:
        200:
          description: Order details
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    $ref: '#/components/schemas/Order'
        404:
          description: Order not found

  # ============= ADMIN USER MANAGEMENT ENDPOINTS =============
...
  /orders:
    post:
      tags: [Orders]
      summary: Create new order from cart
      description: Creates an order from the user's cart and clears the cart.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - shippingAddress
              properties:
                shippingAddress:
                  $ref: '#/components/schemas/Address'
                billingAddress:
                  $ref: '#/components/schemas/Address'
                paymentMethod:
                  type: string
                  enum: [razorpay]
                  default: razorpay
      responses:
        201:
          description: Order created with payment details
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
                    properties:
                      order:
                        $ref: '#/components/schemas/Order'
                      payment:
                        type: object
                        properties:
                          razorpayOrderId:
                            type: string
                          amount:
                            type: number
                          currency:
                            type: string
...
  /orders/{orderId}:
    get:
      tags: [Orders]
      summary: Get order by ID
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: orderId
          required: true
          schema:
            type: string
      responses:
        200:
          description: Order details
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    $ref: '#/components/schemas/Order'
...
  /orders/{orderId}/status:
    put:
      tags: [Orders Admin]
      summary: Update order status with tracking number (Admin only)
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: orderId
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - status
              properties:
                status:
                  type: string
                  enum: [accepted, ready_to_ship, picked_up, in_transit, delivered, cancelled]
                  description: "New order status"
                note:
                  type: string
                  description: "Optional note about the status change"
                trackingNumber:
                  type: string
                  description: "Tracking number (required for shipping statuses: ready_to_ship, picked_up, in_transit)"
                  example: "FDX123456789"
            examples:
              accepted:
                summary: Accept order
                value:
                  status: "accepted"
                  note: "Order confirmed and processing"
              
              ready_to_ship:
                summary: Mark as ready to ship
                value:
                  status: "ready_to_ship"
                  note: "Package ready for courier pickup"
                  trackingNumber: "FDX123456789"
              
              in_transit:
                summary: Mark as in transit
                value:
                  status: "in_transit"
                  note: "Order shipped via FedEx"
                  trackingNumber: "FDX123456789"
      responses:
        200:
          description: Order status updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    $ref: '#/components/schemas/Order'
        400:
          description: Tracking number required for shipping status
        404:
          description: Order not found

  # ============= LAYOUT ENDPOINTS =============
...